Items system and bullets

+++ core/src/main/java/me/partlysunny/shapewars/world/components/player/equipment/item/HasUses.java
@@ -4,6 +4,6 @@ public interface HasUses {
 
     int maxUses();
 
-    long usesRegenRate();
+    float usesRegenRate();
 
 }
